fix: auto logout user if token is expired
